About Bill Burger
Project is Based on billing of the Burger Shop
It follow as tha burger shop has three types of burger and they are as:
The Basic Burger which has the option to add the burger with an addition of other stuff with their price. And Additional of four items can be added to the that burger

The Healthy burger has an option to add two more items than those of the items in the Basic Burger with and additional price

The Delux Burger has the no addtional rather than coke and chips are added as a combi meal .

The All the Addtionals are addded and their complete price is displayed.